backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
pyautogen==0.2.0
openai==1.12.0
python-multipart==0.0.9
pandas==2.2.0
numpy==1.26.3
aiofiles==23.2.1
httpx==0.26.0
pytest==8.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
kaggle==1.5.16 